src/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
requests
python-multipart
boto3>=1.28.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pydantic>=2.4.2